Moved correctly working code to index and depracated all other unused files.
