+++ dep_new_flair.py
